requirements.txt
boto3==1.33.6
cryptography==41.0.7
Werkzeug==2.2.2
Flask==2.1.3
Flask_JWT_Extended==4.6.0
flet==0.16.0
pycryptodome==3.19.0
pyperclip==1.8.2
Requests==2.31.0
